+++ src/init.c
@@ -5,6 +5,7 @@
 
 extern SEXP parse_from_str_ (SEXP str_     , SEXP parse_opts_);
 extern SEXP parse_from_file_(SEXP filename_, SEXP parse_opts_);
+extern SEXP parse_from_raw_ (SEXP filename_, SEXP parse_opts_);
 
 extern SEXP serialize_to_str_ (SEXP x_,                 SEXP serialize_opts_);
 extern SEXP serialize_to_file_(SEXP x_, SEXP filename_, SEXP serialize_opts_);
@@ -18,6 +19,7 @@ static const R_CallMethodDef CEntries[] = {
   
   {"parse_from_str_"  , (DL_FUNC) &parse_from_str_ , 2},
   {"parse_from_file_" , (DL_FUNC) &parse_from_file_, 2},
+  {"parse_from_raw_"  , (DL_FUNC) &parse_from_raw_ , 2},
   
   {"parse_ndjson_file_as_df_"  , (DL_FUNC) &parse_ndjson_file_as_df_  , 5},
   {"parse_ndjson_file_as_list_", (DL_FUNC) &parse_ndjson_file_as_list_, 4},